Move copied `run_entry_point`
